+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Movim Changelog
 v0.9.1 (trunk)
 ---------------------------
  * CSS fixes
+ * Add Last Message Edition support
+ * Improve Post discovery in the News page
 
 v0.9
 ---------------------------